refactor: classes UsuarioSerializer/UsuarioViewSet de serializers e views em tarefas realocadas para serializers e views dentro de usuarios
