+++ tarefas/views.py
@@ -1,15 +1,7 @@
-from django.contrib.auth import get_user_model
 from rest_framework import viewsets
 from rest_framework.permissions import IsAuthenticated
 from .models import Tarefa
-from .serializers import TarefaSerializer, UsuarioSerializer
-
-getModelUsuario = get_user_model()
-
-class UsuarioViewSet(viewsets.ModelViewSet):
-    queryset = getModelUsuario.objects.all().order_by('-date_joined')
-    serializer_class = UsuarioSerializer
-    permission_classes = [IsAuthenticated]
+from .serializers import TarefaSerializer
 
 class TarefaViewSet(viewsets.ModelViewSet):
     serializer_class = TarefaSerializer
