+++ usuarios/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from .models import Usuario
+
+class UsuarioSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = Usuario
+        fields = ['username', 'email', 'url']
