+++ usuarios/views.py
@@ -1,3 +1,13 @@
-from django.shortcuts import render
+from rest_framework import viewsets
+from rest_framework.permissions import IsAuthenticated
+from .models import Usuario
+from .serializers import UsuarioSerializer
+from django.contrib.auth import get_user_model
 
-# Create your views here.
+getModelUsuario = get_user_model()
+
+class UsuarioViewSet(viewsets.ModelViewSet):
+    #queryset = Usuario.objects.all().order_by('-date_joined')
+    queryset = getModelUsuario.objects.all().order_by('-date_joined')
+    serializer_class = UsuarioSerializer
+    permission_classes = [IsAuthenticated]
